Rename changelog.py
